delete a comment

+++ controllers/posts_controller.js
@@ -36,6 +36,16 @@ module.exports.destroy = async function(req, res){
 
         if (post.user == req.user.id){
             post.remove();
+
+            if(req.xhr){
+                return res.status(200).json({
+                    data: {
+                        post_id: req.params.id
+                    },
+                    message: "Post deleted"
+                });
+            }
+
             req.flash('success','Post and associative comments deleted!');
 
             await Comment.deleteMany({post: req.params.id});
